feat: :sparkles: Triggers, dim_metas, and cleaning.
It was added the triggers which will be used for counting the number of unique data in certain columns, it also was added the fact_metas and, it was clean the structure of the query, even fixing some mistakes because of the ETL

+++ DimensionalModel/dim_cliente.sql
@@ -1,5 +1,5 @@
 CREATE TABLE dim_cliente(
-    sk_dim_cliente DOUBLE PRECISION PRIMARY KEY,
+    sk_dim_cliente INTEGER PRIMARY KEY,
     cod_cliente INTEGER NOT NULL UNIQUE,
     nb_cliente VARCHAR(25) NOT NULL,
     ci_rif VARCHAR(45) UNIQUE NOT NULL ,
