+++ DimensionalModel/dim_producto.sql
@@ -1,5 +1,5 @@
 CREATE TABLE dim_producto(
-    sk_dim_producto DOUBLE PRECISION PRIMARY KEY,
+    sk_dim_producto INTEGER PRIMARY KEY,
     cod_producto INTEGER NOT NULL UNIQUE,
     nb_producto VARCHAR(45) NOT NULL,
     descrip_producto VARCHAR(100) NOT NULL,
