+++ DimensionalModel/dim_siniestro.sql
@@ -1,5 +1,5 @@
 CREATE TABLE dim_siniestro(
-    sk_dim_siniestro DOUBLE PRECISION PRIMARY KEY,
+    sk_dim_siniestro INTEGER PRIMARY KEY,
     nro_siniestro INTEGER UNIQUE NOT NULL,
     descrip_siniestro VARCHAR(50) NOT NULL
 );
