+++ DimensionalModel/dim_sucursal.sql
@@ -1,5 +1,5 @@
 CREATE TABLE dim_sucursal(
-    sk_dim_sucursal DOUBLE PRECISION PRIMARY KEY,
+    sk_dim_sucursal INTEGER PRIMARY KEY,
     cod_sucursal INTEGER NOT NULL UNIQUE,
     nb_sucursal VARCHAR(30) NOT NULL,
     cod_ciudad INTEGER NOT NULL,
@@ -7,4 +7,5 @@ CREATE TABLE dim_sucursal(
     cod_pais INTEGER NOT NULL,
     nb_pais VARCHAR(45) NOT NULL
 );
+
